+++ mayan/apps/events/tests/test_event_prune_backends_latest.py
@@ -1,0 +1,190 @@
+from actstream.models import Action
+
+from mayan.apps.testing.tests.base import BaseTestCase
+
+from ..event_prune_backends import (
+    EventLogPruneBackendLatest, EventLogPruneBackendLatestPerObject,
+    EventLogPruneBackendLatestPerObjectEventType
+)
+
+from .mixins.event_prune_backend_mixins import EventLogPruneBackendTestMixin
+
+
+class EventLogPruneBackendLatestTestCase(
+    EventLogPruneBackendTestMixin, BaseTestCase
+):
+    _test_event_prune_backend_class = EventLogPruneBackendLatest
+    _test_event_prune_backend_kwargs = {'number': 3}
+
+    def test_backend(self):
+        self.assertEqual(
+            Action.objects.count(), self._test_event_count - 9
+        )
+
+        self.assertEqual(
+            self._test_object_list[0].target_actions.count(),
+            self._test_object_0_event_count - 6
+        )
+
+        self.assertFalse(
+            Action.objects.filter(pk=self._test_event_list[0].pk).exists()
+        )
+        self.assertFalse(
+            Action.objects.filter(pk=self._test_event_list[1].pk).exists()
+        )
+        self.assertFalse(
+            Action.objects.filter(pk=self._test_event_list[2].pk).exists()
+        )
+        self.assertFalse(
+            Action.objects.filter(pk=self._test_event_list[3].pk).exists()
+        )
+        self.assertFalse(
+            Action.objects.filter(pk=self._test_event_list[4].pk).exists()
+        )
+        self.assertFalse(
+            Action.objects.filter(pk=self._test_event_list[5].pk).exists()
+        )
+
+        self.assertEqual(
+            self._test_object_list[1].target_actions.count(),
+            self._test_object_1_event_count - 3
+        )
+
+        self.assertFalse(
+            Action.objects.filter(pk=self._test_event_list[6].pk).exists()
+        )
+        self.assertFalse(
+            Action.objects.filter(pk=self._test_event_list[7].pk).exists()
+        )
+        self.assertFalse(
+            Action.objects.filter(pk=self._test_event_list[8].pk).exists()
+        )
+        self.assertTrue(
+            Action.objects.filter(pk=self._test_event_list[9].pk).exists()
+        )
+        self.assertTrue(
+            Action.objects.filter(pk=self._test_event_list[10].pk).exists()
+        )
+        self.assertTrue(
+            Action.objects.filter(pk=self._test_event_list[11].pk).exists()
+        )
+
+
+class EventLogPruneBackendLatestPerObjectTestCase(
+    EventLogPruneBackendTestMixin, BaseTestCase
+):
+    _test_event_prune_backend_class = EventLogPruneBackendLatestPerObject
+    _test_event_prune_backend_kwargs = {'number': 3}
+
+    def test_backend(self):
+        self.assertEqual(
+            Action.objects.count(), self._test_event_count - 6
+        )
+
+        self.assertEqual(
+            self._test_object_list[0].target_actions.count(),
+            self._test_object_0_event_count - 3
+        )
+
+        self.assertFalse(
+            Action.objects.filter(pk=self._test_event_list[0].pk).exists()
+        )
+        self.assertFalse(
+            Action.objects.filter(pk=self._test_event_list[1].pk).exists()
+        )
+        self.assertFalse(
+            Action.objects.filter(pk=self._test_event_list[2].pk).exists()
+        )
+        self.assertTrue(
+            Action.objects.filter(pk=self._test_event_list[3].pk).exists()
+        )
+        self.assertTrue(
+            Action.objects.filter(pk=self._test_event_list[4].pk).exists()
+        )
+        self.assertTrue(
+            Action.objects.filter(pk=self._test_event_list[5].pk).exists()
+        )
+
+        self.assertEqual(
+            self._test_object_list[1].target_actions.count(),
+            self._test_object_1_event_count - 3
+        )
+
+        self.assertFalse(
+            Action.objects.filter(pk=self._test_event_list[6].pk).exists()
+        )
+        self.assertFalse(
+            Action.objects.filter(pk=self._test_event_list[7].pk).exists()
+        )
+        self.assertFalse(
+            Action.objects.filter(pk=self._test_event_list[8].pk).exists()
+        )
+        self.assertTrue(
+            Action.objects.filter(pk=self._test_event_list[9].pk).exists()
+        )
+        self.assertTrue(
+            Action.objects.filter(pk=self._test_event_list[10].pk).exists()
+        )
+        self.assertTrue(
+            Action.objects.filter(pk=self._test_event_list[11].pk).exists()
+        )
+
+
+class EventLogPruneBackendLatestPerObjectEventTypeTestCase(
+    EventLogPruneBackendTestMixin, BaseTestCase
+):
+    _test_event_prune_backend_class = EventLogPruneBackendLatestPerObjectEventType
+    _test_event_prune_backend_kwargs = {'number': 3}
+
+    def test_backend(self):
+        self.assertEqual(
+            Action.objects.count(), self._test_event_count - 2
+        )
+
+        self.assertEqual(
+            self._test_object_list[0].target_actions.count(),
+            self._test_object_0_event_count - 1
+        )
+
+        self.assertTrue(
+            Action.objects.filter(pk=self._test_event_list[0].pk).exists()
+        )
+        self.assertTrue(
+            Action.objects.filter(pk=self._test_event_list[1].pk).exists()
+        )
+        self.assertFalse(
+            Action.objects.filter(pk=self._test_event_list[2].pk).exists()
+        )
+        self.assertTrue(
+            Action.objects.filter(pk=self._test_event_list[3].pk).exists()
+        )
+        self.assertTrue(
+            Action.objects.filter(pk=self._test_event_list[4].pk).exists()
+        )
+        self.assertTrue(
+            Action.objects.filter(pk=self._test_event_list[5].pk).exists()
+        )
+
+        self.assertEqual(
+            self._test_object_list[1].target_actions.count(),
+            self._test_object_1_event_count - 1
+        )
+
+        self.assertTrue(
+            Action.objects.filter(pk=self._test_event_list[6].pk).exists()
+        )
+        self.assertTrue(
+            Action.objects.filter(pk=self._test_event_list[7].pk).exists()
+        )
+        self.assertFalse(
+            Action.objects.filter(pk=self._test_event_list[8].pk).exists()
+        )
+        self.assertTrue(
+            Action.objects.filter(pk=self._test_event_list[9].pk).exists()
+        )
+        self.assertTrue(
+            Action.objects.filter(pk=self._test_event_list[10].pk).exists()
+        )
+        self.assertTrue(
+            Action.objects.filter(pk=self._test_event_list[11].pk).exists()
+        )